请求的读  写

+++ src/main/java/org/qiunet/handler/annotation/RequestHandler.java
@@ -17,5 +17,5 @@ public @interface RequestHandler {
 	 * request Command Id
 	 * @return
 	 */
-	int ID();
+	short ID();
 }
